requirements.txt
requests
pandas
scipy
matplotlib
seaborn
python-dotenv